src/main/java/it/unipd/mtss/model/EItem.java
////////////////////////////////////////////////////////////////////
// Marco Marchiante 1222397
// Michele Bettin 1216735
////////////////////////////////////////////////////////////////////

package it.unipd.mtss.model;

public class EItem {
    
    ItemType itemType;
    String name;
    double price;

    public EItem(ItemType type, String name, double price) {
        this.itemType = type;
        this.name = name;
        this.price = price;
    }

    public ItemType getItemType() {
        return itemType;
    }

    public String getName() {
        return name;
    }

    public double getPrice() {
        return price;
    }

}
